RSE-905: Add Upgrader for Adding Case Category To CG Extends. Removed Un-used Setup.
