+++ CRM/Civicase/Setup/AddCaseCategoryCgExtendsValue.php
@@ -10,12 +10,12 @@ class CRM_Civicase_Setup_AddCaseCategoryCgExtendsValue {
   const CASE_CATEGORY_LABEL = 'Case (Cases)';
 
   /**
-   * Add Prospecting as a valid Entity that a custom group can extend.
+   * Add Cases as a valid Entity that a custom group can extend.
    */
   public function apply() {
     $result = civicrm_api3('OptionValue', 'get', [
       'option_group_id' => 'cg_extend_objects',
-      'label' => self::CASE_CATEGORY_LABEL,
+      'value' => CaseCategoryHelper::CASE_TYPE_CATEGORY_NAME,
     ]);
 
     if ($result['count'] > 0) {
